integrated tables with backend API
