+++ src/table_builder/utilities.js
